+++ @xen-orchestra/lite/CHANGELOG.md
@@ -16,6 +16,7 @@
 - [Host/Header] Add master host icon on host header (PR [#8512](https://github.com/vatesfr/xen-orchestra/pull/8512))
 - [Host] Add dashboard view (PR [#8504](https://github.com/vatesfr/xen-orchestra/pull/8504))
 - [VM/New] Add topology field in VM creation (PR [#8491](https://github.com/vatesfr/xen-orchestra/pull/8491))
+- [VM] Improved rooting of default vm page (PR [#8553](https://github.com/vatesfr/xen-orchestra/pull/8553))
 
 ## **0.9.1** (2025-04-02)